Split the step definition file into 3 pieces. The setup, assertions and request-response step definitions.
